cleaning things up
adding error checking and more output

+++ samples/EFSBackup/efs-backup.sh
@@ -21,18 +21,37 @@ retain=$4
 efsid=$5
 
 # Prepare system for rsync
-#echo 'sudo yum -y update'
-#sudo yum -y update
+echo '*********  Updating Systems and Installing Utils *********'
+echo 'sudo yum -y update'
+sudo yum -y update
 echo 'sudo yum -y install nfs-utils'
 sudo yum -y install nfs-utils
-echo 'sudo mkdir /backup'
-sudo mkdir /backup
-echo 'sudo mkdir /mnt/backups'
-sudo mkdir /mnt/backups
+
+
+
+echo '\n\n'
+echo '*********  Building Mount Points  *********'
+if [[ ! -d /backup ]]; then
+  echo 'sudo mkdir /backup'
+  sudo mkdir /backup
+fi
+
+if [[ ! -d /mnt/backups ]]; then
+  echo 'sudo mkdir /mnt/backups'
+  sudo mkdir /mnt/backups
+fi
+
+echo '\n\n'
+echo '********* Mounting EFS Drives to Mount Points  *********'
 echo "sudo mount -t nfs -o nfsvers=4.1 -o rsize=1048576 -o wsize=1048576 -o timeo=600 -o retrans=2 -o hard $source /backup"
-sudo mount -t nfs -o nfsvers=4.1 -o rsize=1048576 -o wsize=1048576 -o timeo=600 -o retrans=2 -o hard $source /backup
+sudo mount -t nfs -o nfsvers=4.1 -o rsize=1048576 -o wsize=1048576 -o timeo=600 -o retrans=2 -o hard $source /backup || { echo "mount /backup FAILED"; exit $?; }
+
 echo "sudo mount -t nfs -o nfsvers=4.1 -o rsize=1048576 -o wsize=1048576 -o timeo=600 -o retrans=2 -o hard $destination /mnt/backups"
-sudo mount -t nfs -o nfsvers=4.1 -o rsize=1048576 -o wsize=1048576 -o timeo=600 -o retrans=2 -o hard $destination /mnt/backups
+sudo mount -t nfs -o nfsvers=4.1 -o rsize=1048576 -o wsize=1048576 -o timeo=600 -o retrans=2 -o hard $destination /mnt/backups || { echo "mount /mnt/backups FAILED"; exit $?; }
+
+
+echo '\n\n'
+echo '********* Removing OLD Backups  *********'
 
 # we need to decrement retain because we start counting with 0 and we need to remove the oldest backup
 let "retain=$retain-1"
@@ -41,6 +60,8 @@ if sudo test -d /mnt/backups/$efsid/$interval.$retain; then
   sudo rm -rf /mnt/backups/$efsid/$interval.$retain
 fi
 
+echo '\n\n'
+echo '********* Shifiting Backups  *********'
 
 # Rotate all previous backups (except the first one), up one level
 for x in `seq $retain -1 2`; do
@@ -67,8 +88,13 @@ if [ ! -d /mnt/backups/efsbackup-logs ]; then
   echo "sudo chmod 700 /mnt/backups/efsbackup-logs"
   sudo chmod 700 /mnt/backups/efsbackup-logs
 fi
-echo "sudo rm /tmp/efs-backup.log"
-sudo rm /tmp/efs-backup.log
+
+if [ -f /tmp/efs-backup.log ]; then
+  echo "sudo rm /tmp/efs-backup.log"
+  sudo rm /tmp/efs-backup.log
+fi
+
+
 echo "sudo rsync -ah --stats --delete --numeric-ids --log-file=/tmp/efs-backup.log /backup/ /mnt/backups/$efsid/$interval.0/"
 sudo rsync -ah --stats --delete --numeric-ids --log-file=/tmp/efs-backup.log /backup/ /mnt/backups/$efsid/$interval.0/
 rsyncStatus=$?
